# openapi/scintx.openapi.yaml
openapi: 3.1.0
info:
  title: SCINTX — Supply Chain INTelligence Exchange
  version: 1.0.0
  summary: Vendor-neutral interoperability API for security scanners.
  description: |
    SCINTX sits between consumers (registries, enterprises) and security providers,
    defining a thin canonical envelope around existing standards (PURL, OSV,
    CycloneDX, SPDX, SARIF, CSAF, in-toto, SLSA, CloudEvents). Normative data
    schemas are standalone JSON Schema Draft 2020-12 documents in /schema;
    this OpenAPI document describes the HTTP surface and references them.
  license:
    name: Eclipse Public License 2.0
    url: https://www.eclipse.org/legal/epl-2.0/
  contact:
    name: SCINTX
    url: https://github.com/yeeth-security/scintx

servers:
  - url: https://scintx.example
    description: Production
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Submissions
    description: Create and inspect security analysis submissions.
  - name: Providers
    description: Discover providers and their capability manifests.
  - name: Artifacts
    description: Upload artifact content for scanners that require exact bytes.
  - name: Discovery
    description: Service discovery and well-known endpoints.
  - name: Events
    description: CloudEvents emitted during the submission lifecycle.

security:
  - HttpMessageSignatures: []

paths:
  /v1/submissions:
    post:
      tags: [Submissions]
      operationId: createSubmission
      summary: Submit an artifact for security analysis
      description: |
        Creates a submission. Returns 202 Accepted with the initial Submission
        state. Provider invocation proceeds asynchronously; the client polls
        GET /v1/submissions/{id} or receives CloudEvents.

        Idempotency: if an Idempotency-Key header is supplied, a second request
        with the same key and the same canonical body returns the original
        submission. Same key with a different body returns 409 Conflict.
        No exactly-once guarantee.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schema/submission-create.schema.json'
            examples:
              purlVulnerability:
                value:
                  schema_version: "1.0.0"
                  artifact:
                    purl: "pkg:pypi/requests@2.32.3"
                  requested_capabilities: ["vulnerability"]
                  policy_ref: "registry-default"
              findingsOnly:
                value:
                  schema_version: "1.0.0"
                  artifact:
                    purl: "pkg:npm/left-pad@1.3.0"
                  requested_capabilities: ["vulnerability"]
      responses:
        '202':
          description: Submission accepted
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
              description: URL of the created submission.
          content:
            application/json:
              schema:
                $ref: '../schema/submission.schema.json'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'

  /v1/submissions/{id}:
    get:
      tags: [Submissions]
      operationId: getSubmission
      summary: Get submission state
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: Submission state
          content:
            application/json:
              schema:
                $ref: '../schema/submission.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/submissions/{id}/results:
    get:
      tags: [Submissions]
      operationId: listSubmissionResults
      summary: List provider results for a submission
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: Provider results
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: '../schema/provider-result.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/results/{id}:
    get:
      tags: [Submissions]
      operationId: getResult
      summary: Get a provider result by id
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Provider result
          content:
            application/json:
              schema:
                $ref: '../schema/provider-result.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/decisions/{id}:
    get:
      tags: [Submissions]
      operationId: getDecision
      summary: Get a policy decision by id
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Policy decision
          content:
            application/json:
              schema:
                $ref: '../schema/policy-decision.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/submissions/{id}/resume:
    post:
      tags: [Submissions]
      operationId: resumeSubmission
      summary: Resume a deferred submission
      description: |
        Transitions a deferred submission back to running. Only valid when
        status=deferred. Returns the updated submission state.
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: Submission resumed
          content:
            application/json:
              schema:
                $ref: '../schema/submission.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/submissions/{id}/adjudicate:
    post:
      tags: [Submissions]
      operationId: adjudicateSubmission
      summary: Share a consumer final resolution back to the gateway
      description: |
        Resolution happens wherever the consumer plugs SCINTX results
        (registry UI, ticketing, CI gate, etc.). This endpoint records that
        shared allow/deny outcome on the gateway for audit and webhooks.

        The prior machine PolicyDecision is kept immutable. A new decision is
        stored with extensions.origin=consumer and prior_decision_id, and
        submission.decision_id is updated to the new id. Emits
        org.eclipse.scintx.policy-decision.resolved.v1.
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schema/adjudicate-request.schema.json'
            examples:
              allowFromRegistryUI:
                value:
                  decision: allow
                  actor: alice@example.com
                  source: registry-ui
                  rationale: Accepted risk until lodash>=4.17.21
      responses:
        '201':
          description: Consumer resolution recorded
          content:
            application/json:
              schema:
                $ref: '../schema/policy-decision.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /v1/providers:
    get:
      tags: [Providers]
      operationId: listProviders
      summary: List registered providers with capability summaries
      responses:
        '200':
          description: Provider list
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      type: object
                      required: [id, name, capabilities]
                      properties:
                        id: {type: string}
                        name: {type: string}
                        capabilities:
                          type: array
                          items: {type: string}
                          description: Capability IDs with versions (e.g. vulnerability:v1).

  /v1/providers/{id}/capabilities:
    get:
      tags: [Providers]
      operationId: getProviderCapabilities
      summary: Get a provider's capability manifest
      parameters:
        - $ref: '#/components/parameters/ProviderId'
      responses:
        '200':
          description: Capability manifest
          content:
            application/json:
              schema:
                $ref: '../schema/provider-capabilities.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/artifacts:
    post:
      tags: [Artifacts]
      operationId: uploadArtifact
      summary: Upload artifact content
      description: |
        Uploads artifact content and returns a content-addressed reference
        with computed SHA-256 digest. Use this to obtain a urn:scintx:blob:...
        URI for providers that require exact bytes (malware, static analysis).
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Artifact stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  artifact_ref:
                    $ref: '../schema/artifact-ref.schema.json'

  /v1/artifacts/{digest}:
    head:
      tags: [Artifacts]
      operationId: checkArtifact
      summary: Check artifact existence
      description: Returns 200 if the artifact with the given digest exists, 404 otherwise. No body.
      parameters:
        - name: digest
          in: path
          required: true
          schema:
            type: string
            pattern: '^sha256:[0-9a-f]{64}$'
      responses:
        '200':
          description: Artifact exists
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/.well-known/scintx:
    get:
      tags: [Discovery]
      operationId: wellKnown
      summary: Service discovery
      responses:
        '200':
          description: Service metadata
          content:
            application/json:
              schema:
                type: object
                required: [standard, version, endpoints, auth_profiles]
                properties:
                  standard: {type: string, const: scintx}
                  version: {type: string}
                  endpoints:
                    type: array
                    items: {type: string}
                  auth_profiles:
                    type: array
                    items: {type: string}
                    description: Required-to-implement auth profile is rfc9421.
                  event_types:
                    type: array
                    items: {type: string}

  /v1/events:
    get:
      tags: [Events]
      operationId: listEvents
      summary: List emitted CloudEvents
      description: |
        Returns the in-process event log for debugging/conformance.
        Production clients should subscribe via signed webhooks
        (SCINTX_WEBHOOK_URL): CloudEvents POST with Content-Digest (RFC 9530)
        and X-Scintx-Signature HMAC. Full inbound RFC 9421 API auth is a
        separate profile.
      parameters:
        - name: subject
          in: query
          schema: {type: string}
          description: Filter by event subject (submission id).
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '../schema/event.schema.json'

components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string}
      description: |
        scoped to the authenticated caller. Same key + same canonical request
        returns the original submission. Same key + different body returns 409.
    SubmissionId:
      name: id
      in: path
      required: true
      schema: {type: string}
    ProviderId:
      name: id
      in: path
      required: true
      schema: {type: string}

  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema:
            $ref: '../schema/problem-details.schema.json'
    Unprocessable:
      description: Request was well-formed but semantically invalid (e.g. artifact has no valid identity)
      content:
        application/problem+json:
          schema:
            $ref: '../schema/problem-details.schema.json'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '../schema/problem-details.schema.json'
    Conflict:
      description: Conflict with existing state (e.g. idempotency key mismatch, resume on non-deferred submission)
      content:
        application/problem+json:
          schema:
            $ref: '../schema/problem-details.schema.json'

  securitySchemes:
    HttpMessageSignatures:
      type: apiKey
      in: header
      name: Signature
      description: |
        RFC 9421 HTTP Message Signatures. Required-to-implement auth baseline
        for all SCINTX API endpoints. The Signature-Input and Signature headers
        carry signed HTTP components (method, target-uri, content-digest,
        content-type) with created/expires/keyid. OAuth 2.1 bearer is an
        optional second profile.